// src/types/finnhub.ts
export interface SearchResult {
  description: string;
  displaySymbol: string;
  symbol: string;
  type: string;
}

export interface SearchResponse {
  count: number;
  result: SearchResult[];
}

export interface Quote {
  c: number;
  d: number;
  dp: number;
  h: number;
  l: number;
  o: number;
  pc: number;
  t: number;
}

export interface StockData {
  symbol: string;
  companyName: string;
  currentPrice: number;
  change: number;
  percentChange: number;
}